agrisage-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.0
numpy==1.24.3
requests==2.31.0
twilio==8.10.0